+++ walle/api/space.py
@@ -37,7 +37,7 @@ class SpaceAPI(SecurityResource):
         """
         page = int(request.args.get('page', 0))
         page = page - 1 if page else 0
-        size = float(request.args.get('size', 10))
+        size = int(request.args.get('size', 10))
         kw = request.values.get('kw', '')
 
         space_model = SpaceModel()
