+++ walle/api/task.py
@@ -36,7 +36,7 @@ class TaskAPI(SecurityResource):
         """
         page = int(request.args.get('page', 0))
         page = page - 1 if page else 0
-        size = float(request.args.get('size', 10))
+        size = int(request.args.get('size', 10))
         kw = request.values.get('kw', '')
 
         task_model = TaskModel()
